alert-bad-processes: alert in case something uses 100% of CPU
Put this script into crontab, e.g.

*/2 * * * * root test -f /root/bin/alert-bad-processes && /root/bin/alert-bad-processes 0 1 &> /dev/null

A Linux script which could alert you (visually or/and audibly) when there is/are process(es) which use(s) 100% of CPU.

It's written specifically for X.org. It's currently fine-tuned for LXDM but there's a universal version which is suitable if `w` works for you.