Show a message from root in the user X session
I'm probably inventing a bicycle here since something like that must be possible but here it is anyways.

You can use this script to send messages to a currently logged in user using xmessage or notify-send. This script is hard coded for X.org and most likely won't work with Wayland. It's pretty self-explanatory if you run it without arguments.